refactor: split the credentialed-MCP per-request resolution out of its controller
The controller crossed its tier cap once the context became deferred. Its
per-request resolution moves to a sibling: the controller is the transport, and
`_request_context` is what a request resolves from application state.

The unit-api conftest rebound `synthorg.api.app.create_app` to the
kwarg-flattening test helper. That rebinding is process-global and outlives the
package's collection, so any other module that boots the app for its own reasons
was handed a different signature depending on which suites were collected
alongside it. The module attribute now keeps the real signature and carries only
the typeguard suppression it was there for.
